refactor: change folder structure and naming
